pyproject.toml
[project]
name = "scraper-server"
version = "0.1.0"
description = "FastAPI server for scraper platform"
requires-python = ">=3.10,<3.13"
dependencies = [
    "azure-monitor-opentelemetry>=1.8.1",
    "azure-servicebus>=7.14.2",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "pydantic>=2.11.4",
    "redis>=5.3.0",
    "requests>=2.32.3",
    "opentelemetry-instrumentation-fastapi>=0.58b0",
    "python-dotenv>=1.2.1",
]
